setting、about调整
