--- app/src/main/assets/termOfService_en.html
@@ -1,24 +1,0 @@
-<p>You must accept the following term of service to continue using BookShlef:</p>
-
-<h3>1. Copyright</h3>
-
-<ul>
-<li>The copyright of the book information got from web services are reserved by these serives. BookShelf is not responsible for copyright issues. If you want to use these data, you should ask these services for permission.</li>
-<li>You should not use BookShelf to do anything violating copyright rules。</li>
-</ul>
-
-<h3>2. Information Collection</h3>
-
-<p>BookShelf is an open-source,free app, we won't collect your personal information such as your book list, bookshelves, labels, etc. However, to make this app better, we will collect the following data:</p>
-
-<ul>
-<li>Crash Report of BookShelf</li>
-<li>The usage frequency of each function BookShelf provided</li>
-<li>Basic information about your device (only include Android Version, Device Name etc, EXCLUDE any personal information like IMEI, MAC, Android ID etc)</li>
-</ul>
-
-<p>In fact, above are the functions provided by Fabric Answers and Crashlytics plugins, we promise that these data is only used to improve user experience, we won't and cannot use them to track back to any specific user. Also, we won't share these data with any third-party people or organizations.</p>
-
-<h3>3. Others</h3>
-
-<p>Other things not mentioned above are subject to relevant laws and rules. If you have any questions about that, you can <a href="mailto:smartjinyu@gmail.com">contact developers</a> or view BookShelf's source code on <a href="https://github.com/smartjinyu/MyBookshelf">GitHub</a>.</p>
