--- app/src/main/assets/termOfService_zh.html
@@ -1,25 +1,0 @@
-<p>继续使用 BookShelf(以下简称“本软件”) 视为同意以下条款：</p>
-
-<h3>1.版权问题</h3>
-
-<ul>
-<li>本软件从网络服务中获取的书籍信息，其版权约束按照服务商的相关说明规定，与本软件无关，如需商业应用，请联系相关服务商获得许可。</li>
-<li>请勿使用本软件进行违反版权法规的活动，包括但不限于侵犯出版社、原作者权利以及书籍信息提供网站权利的活动。本软件不对上述活动负责。</li>
-</ul>
-
-<h3>2.信息收集</h3>
-
-<p>本软件开源免费，不会上传您的书籍列表、书架、标签等隐私数据，但是为了更好地改进产品体验，我们会收集以下信息：</p>
-
-<ul>
-<li>本程序的崩溃报告</li>
-<li>本程序各项功能的使用次数</li>
-<li>您的设备基本信息（仅包括 Android Version、Device Name 等，不包括 Android ID、MAC、IMEI 等隐私数据）</li>
-</ul>
-
-<p>事实上，以上功能是由 Fabric.io 的 Answers 和 Crashlytics 插件提供，我们承诺，以上数据仅用于改进用户体验，不会用于也无法定位具体用户，且不会与第三方分享。</p>
-
-<h3>3.其他</h3>
-
-<p>其他未尽事宜请参看相关法律法规（主要是懒得写太多，写了也没人看）。如有问题，欢迎<a href="mailto:smartjinyu@gmail.com">联系作者</a>或前往 <a href="https://github.com/smartjinyu/MyBookshelf">GitHub</a> 查看本软件源代码。</p>
-
